Moves css files into styles directory
