+++ stop-watch/src/styles/Controls.css
