+++ stop-watch/src/styles/Laps.css
