+++ stop-watch/src/styles/MainTimer.css
