+++ stop-watch/src/styles/index.css
